[436343] Add all refactor menus after the Edit group
The diagram split introduced a contributed refactoring menu which must
not appear between the new child categorized menus.
AbstractMenuBuilder.getLabel() is now public, it allows
CustomSiriusActionBarContributor to use the label to get contributed
refactor menus and place them after the edit group.
It will also allow to regroup the menu builders with the same label and
to sort the resulting menu by labels in future.

Bug: 436343
Change-Id: I2bfd84bf8037ff90d6282b42a41999c64b1dcbb9
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>