[436465] Complete the variable support in service: and var: interpreters
Solution that uses the VariableManager (as query legacy, compound
interpreter, ..) to manage the push/peak/pop mechanism for the
variables.

The unSetVariable method was not implemented. Now it uses a
VariableManager with its the variable value stacks: when a variable is
unset, the next getValue, will return its previous value.

Bug: 436465
Change-Id: I72333bd0e636cb08e731f4858703784f3063e8c8
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>