[test] Workaround for Luna-specific behavior
It looks like starting from Luna, if an editor has a title starting with
a space, getTitle() will return a trimmed version without the leading
space. This can break some tests which expected the editor's title to
match the representation's name, when the sample data used for the test
create a name starting with a space (typically a diagram named " package
entities" because the sample ecore package used by the test has no
name).

Change-Id: I9be2ef919f94d800704f96e5c454f1940f9c69de
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>