[435507] Fix border node location
During this feature, a bug concerning border node location has been
discovered. This commit fix this problem by changing the
CanonicalDBorderItemLocator (for GMF location) and the
DBorderItemLocator (for draw2d location). This bug concerns border node
located on the border (like when snapToGrid is enabled and the grid is
on a border of the parent).

Bug: 435507
Change-Id: Id227fbd0876c0fb1ffa4bd2ba5ce86fb947c2218
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>