[438691] Add "touched" mode for selection with rectangle
Currently, you can select several elements by using a selection
rectangle. To be selected, an element must be completely contained by
the selection rectangle.

This commit adds a new mode ("touched" mode), in which to be selected,
an element must intersect the selection rectangle.

This new mode will be activated when the user selects elements from
right to left. The current mode remains when the user selects elements
from left to right.

Bug: 438691
Change-Id: I3f376bc9b2292cf56835338c8a3ae9a0140bc74d
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>