[cleanup] Avoid potential NPE in ViewEdgeFigure
AbstractDiagramEdgeEditPart.ViewEdgeFigure has 2 methods (paintFigure
and paintChildren) which try to access the Sirius model (the DEdge) in
order to decide if it is visible (and thus should be painted or not). In
some circumstances, a Draw2D deferred layout can trigger these methods
in a state where the models and edit parts are invalid, and in that
situation, the call to resolveSemanticElement() breaks with an NPE in
TransactionImpl.commit().

The fact that ViewEdgeFigure (from the Draw2D layer) has access to the
edit part and tries to read the model is clearly broken in itself. In
the case of paintFigure and paintChilren, it is probably not even needed
because the visibility is normally propagated the other way around (i.e.
from the DEdge model element towards the Draw2D figure). I'll open a
ticket to remember to clean this up in the future, but in order to
minimize the changes in the context of a maintenance branch the guard
which is already present in the layout() method will do:

    if (!isActive()) {
         return;
     }

Change-Id: I1f79c2dae650593e10e211a383421caec8e5faa8
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>