[439871] Move the diagram refresh logic down to the non-ui plug-in
Move the DiagramDialect and all the code it depends on into the
core (non-ui) diagram plug-in.

The CanonicalSynchronizer implementation (DDiagramCanonicalSynchronizer)
is still defined in the ui, but invoked from the core through the new
canonicalSynchronizerFactoryOverride extension point.

Also move the ViewpointSelectionCallback logic in a dedicated class in
the core plug-in, delegating from the old one.

Bug: 439871
Change-Id: I06f70d0264231ef364163f42cfbb49d2e6055bdb
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>