[439791] Force mainRepresentationFileURI computation after its creation
The current project on which we are adding the ModelingProject nature
does not contain any representation file, so it has been marked as
invalid and has no mainRepresentationFileURI and hence no session. So just
after creating this main aird, we now recompute (force) the
mainRepresentationURI and the valid state.

Cherry-picked from 436375

Bug: 439791
Change-Id: Idb52a549369d646434f0c483912f47c5beab5a24
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>
Cherry-picked-from: 62ceab88fb2243c7e9ce681150dc0fade84d3921