[439338] Avoid NPE during Tabbar items visibility changes
When the user change the selection from a diagram element to the diagram
(or from the diagram to the diagram element) and if he has made one of
the tabbar menu managers visible, when the ToolbarManager is updated it
detects that the menu are now non visible: all their children menu items
are not visible anymore, their corresponding action are selection
listeners and update themselves when the selection changes). The
ToolbarManager decides to remove them (and dispose them) but when the
selection is changed back, it decides to make the element visible again,
the manager was disposed before, so the page field is null and we get an
NPE. 

For the other items (which are not menu managers), the visibility state
is mainly computed from the visibleWhen expressions but for menu
managers and their actions which comes from GMF, the first visibility
refresh from the selection is done first.

This field is not required as when the menu is set to visible, the page
is the current active page and we can directly get it.

Bug: 439338
Change-Id: I65bad0bb4205850cc309a45402e49b93f13c4bdb
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>