[427871] Remove calls to SelectionCommandAppender
The current bug aims to delete the IChangeListener installed by each
DDiagramEditorImpl on its DDiagram. The SelectionCommandAppender is used
to decorate several commands in order to select the newly created
elements during the decorated command execution. The
SelectionCommandAppender uses the IChangeListener to detect those
elements to select. 

This commit removes calls to SelectionCommandAppender and the add/remove
of the IChangeListener during the DDiagramEditorImpl initialization. 

The next commit should delete the SelectionCommandAppender and replace
it by a more efficient and generic mechanism.

Bug: 427871
Change-Id: I15407c3e5f08c1f111d80639be332d57ef0b314b
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>