[437393] Improve labels in layers creation menus
Now we do not display the feature name and the type name, the menus
were: "Additional Layer Additional Layer" and "Default Layer Layer". 

Now the menus labels are "Additional Layer" and "Default Layer".

Also disbled the creation of a default layer for a
DiagramImportDescription as the implementation of its getter will only
look for the default layer of the imported diagram description.

Bug: 437393
Change-Id: Iaea18bbd7508c8f4cf8313d5429872561664c74b
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>