[438074] Creation of a Wizard for VSM Initializer.
An initilizer action has been added in the contextual menu of the VSM Editor when a Viewpoint is selected. It launches a wizard which allow to select and configure a pattern. 

The Class Diagram pattern is currently the only available pattern. Its configuration page allows to specify the base id used as prefix for all IdentifiedElement of the created Class Diagram skeleton. The specifier also has the possibility to configure between a single or many delete and direct edit tools.

The Wizard is in a DRAFT state. It still has to be moed in the proper plugin, an extension point or an API has to be written to provide the patterns.

Bug: 438074
Change-Id: I22d1a585323d9d8e8c292d830310d1b08e540265
Signed-off-by: João Martins <joaomartins27396@gmail.com>