[426583] Integrate a first version of the developer doc
This includes almost all of the old documentation, but many parts were
not fully reviewed and may be obsolete/misleading. I'm pushing them
anyway so that they are at least in the repo were there is a chance they
will be improved/fixed over type; we may disable the most problematic
parts in the TOC and top-level developer page before the release to
avoid confusing readers.

Bug: 426583
Change-Id: I4f82afb9f4e1d4efa9450b15a515205e47faaa21
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>