[427871] Minimize selection of the newly created elements.
Selecting all created elements could add a lot of noise in the editor,
this could be annoying for the user. This commit filters the newly
created elements to take only the roots of the created group of
elements: if an element has its parent in the created elements list, it
will not be selected. 

This also prevent additional refresh on Trees.

Bug: 427871
Change-Id: I1297072d2ae9d1dd3e96cc54f13a96c2340433d8
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr