[441417] Avoid same GMF locations after moving several border nodes.
During a move of several border nodes simultaneously, the feedbacks of
these nodes are ignored for conflict detection (in DBorderItemLocator).
This causes that all border nodes have the same GMF location after the
move. This is not acceptable, because this can cause unexpected moves of
border node later.

Add a new list of figures to the DBorderItemLocator.conflicts() method
to allow to complete the brother figures with other figures. In our case
these figures are the feedback figures. Several methods are completed to
have this new parameter.
This list of figures is filled during the showChangeBoundsFeedback()
method (and cleaned during eraseSourceFeedback() method).

Bug: 441417
Change-Id: Icfd8f93a2cb35969d78d319c36fd592f85ac9331
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>