[427869] Simplify DeleteDRepresentationTask
Get the model accessor from the registry only once.
DeleteRepresentationTask is added when deleteReferences == false, the task is created in
DeleteSeveralDDiagramElementsTask. This task will be executed before the
DeleteDRepresentationElementTask created on the same element. So we can
remove the GMFUtil.tearDownOutgoingReferences as its goal has no sense:
when an element is detached/removed from its container, the cross
referencer adapter is notified , the unsetTarget is called and the
inverse references from this object are removed. Furthermore the
DeleteDrepresentationElementTask (or later the
DanglingRefRemovalTrigger) will perform a eDelete on the current
DRepresentation, and the
ECrossReferenceAdatper should then be up to date.

Bug: 427869
Change-Id: I8c4ceaf8a387d83cb10e1611325cabc4c25bf5ca
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>