[427869] Refactor delete representation and representation element tasks
DeleteSeveralDDiagramElementsTask, DeleteTreeElementsTask and
DeleteTableElementsTask were very similar, they have been merged into
DeleteRepresentationElementsTask. Several issues were detected: each
DeleteXXXElementTask was deleting only the representations of their
corresponding Dialect, a lot of code was duplicated and some unnecessary
treatments were done.

We do not need to create the DeleteDRepresentationTask for
DRepresentation elements nor we need to execute the permission check :
the eDelete/eRemove done in DeleteDRepresentationElementTask will check
this and throw the same LockedInstanceException if required.

The diagram behavior has been kept for the vpElements computation: in
manual refresh compute the DSemanticDecorator to delete in the current
representation. For Tree and Table the previous solution was to delete
only the current DRepresentationElement.

Bug: 427869
Change-Id: I4c002057c8bf99c8d78c7341f1eca24068879b62
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>