[427869] Remove DeleteDiagramElementsTriggerOperation mechanism
DeleteDiagramElementsTriggerOperation was used to remove any
DSemanticDecorator which target semantic elements were deleted by the
external action. We do this in a rather inefficient way:
* it uses a ChangeListener which installs a ContentAdapter on the model;
* the ChangeListener keeps in memory all added/removed/modified
elements;
* after the task's execution we use APIs like
TaskHelper.getDElementToClearFromSemanticElements() which iterate on all
the content of the representation models;
* if a tool executes N different Java actions, we do all of this N
times...

In addition, what this task does is inconsistent with what happens for
other tasks (that use pure ModelOperations):
* it does not take into account the "Automatic Refresh" setting (which
is bad);
* it removes representations whose target element has been removed
(which is good, and should probably be generalized).

The IExternalJavaAction2 interface has also been removed: it was
introduced only to allow for implementers to explicitly declare when
their actions do not delete any element so that we can avoid the costly
DeleteDiagramElementsTriggerOperation in these cases.

The DeleteDRepresentationElementsTask has been modified to delete at
least its representation element if the corresponding semantic element
is
null or does not have a container: now that
DeleteDiagramElementsTriggerOperation has been removed and because
SiriusCommand.getDeletedObject() returns an empty list for
ExternalJavaActionTask, this has been done to avoid the red cross
decorator after a delete from model in manual refresh (if the delete
tool use an external java action to do the delete). This a temporary
behavior to minimize the changes for the more simple cases.

Bug: 427869
Change-Id: I32da343309e798cb5363157de4161fce5c20b616
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>