[441483] Source code for automated tests (JUnit & SWTbot)
This is the code accepted as CQ #8550, with a single small change in
plugins/org.eclipse.sirius.sample.interactions.design/doc/license.txt,
as per requested by the IP team (see comment #5 in the CQ).

For reference, this corresponds to the state of the code in the old
Viewpoint repo as of commit f33e43835783f3515b4fb9df1819b9f8ea7b3347
there.

Bug: 441483
Change-Id: I66351a2a699b4db9cdcf3aa32c1026af2f191988
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>