[441770] Cache the previous invocation of StyleConfigurationRegistry.getStyleConfiguration()
StyleConfigurationRegistry.getStyleConfiguration() can be called a lot
in sensitive loops. It used to create a new StyleWrapper instance almost
every time, but StyleWrapper.hashCode(), which is used to find a match
in the registry, can be very costly (it uses EcoreUtil.getURI() to match
elements accross sessions and ResourceSets).

In practice, at least in the case of this bug, successive calls to
getStyleConfiguration() will often use the same mapping instance and
styles with the same StyleDescription many times in a row. Simply
caching the parameters and result of the previous invocation, and
matching the parameters using == gives a performance boost in these
scenario, with very little overhead in time and memory in cases where we
revert to the previous behavior.

Bug: 441770
Change-Id: I991f0ab59c88ca63bd5ea4eefc3a605b0004192c
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>