[443950] Compute the representations to export only when invoked
Since the computation of the representations is only done at execution,
isEnabled() does not have access to the info, and thus always return
true. Thus the action always appear as enabled in the context menu.

Related to this, when the user invokes the action but no representation
is found to export, there is currently a popup explaining that no
exportable representation were found.

To do that, the ExportAsRepresentationAction override the super method
AbstractExportRepresentationsAction.run() to avoid the execution of this
export representation action if there is no representation to export.

Bug: 443950
Change-Id: Ie67a5abb518649cbca63c2464fde768cd1f1886e
Signed-off-by: Belqassim Djafer <belqassim.djafer@obeo.fr>