[427871] Let the editor install its SelectCreatedDRepresentationListener
Keep the dialect agnostic aspect of the post selection listener but
remove the global/centralized listener to let each dialect editor
install or not the listener. 

Diagram is currently the single dialect to use it, as on Tree and Table
the global mechanism leads to performance issues (set selection can
cause expansion, then refresh, then selection, ...)

The retrigger capacity of the selection runnable has been removed, as
the post commit is no more added by the EditingSession (it was installed
before the dialect specific refreshers, and the refresh was eventually
not done before its first runnable execution). 

Bug: 427871
Change-Id: Ib2510f7a930b3d8767fee2e1ee273e8e506ccc83
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>