[442677] Use the crossreferencer in TaskHelper.getDElementToClear...
If we retrieve the semantic cross referencer of the current session, we
use it to compute TaskHelper.getDElementToClearFromSemanticElements else
we fall back to the previous computation with the eAllContents done on
the root parameter.

The previous computation has also been corrected: if the target of a
DRepresentationElement is in semantic elements it must be removed even
if the semantic elements are not empty. Furthermore the target is not
necessarily in those semantic elements. 

Bug: 442677
Change-Id: I8f207e79eb76488ed0edbdb7fe18438f480df6e3
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>