[442677] Postpone some delete subtasks creation at command execution
As in the "delete with tool" case, several delete sub tasks are now
postponed and done during execution of the delete.

The DeleteWithoutToolTask has been introduced to avoid code duplication
and assure to control and maintain the delete without tool of the
different dialects.

The type of collections used to collect the semantic elements to delete
has been changed from hash set to linked hash set, to avoid a random
delete order and improve the reproducibility.

Bug: 442677
Change-Id: I37499334029ba65c41affbcef4a94aebbae55c6b
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>