[442337] Make sure the RepresentationChangeAdapter is installed after the cross-referencer
The RepresentationChangeAdapter triggers a refresh of the UI when
representations are added or removed, but the UI uses
DialectManager.getRepresentations() to find the representations to show.
The new (faster) implementation of that method relies on the
cross-referencer being up-to-date, so we must make sure the
RepresentationChangeAdapter is installed after it for the UI to
correctly reflect the changes.

Bug: 442337
Change-Id: Id2c38330853b52967c0fdf9f4ddb8afb5462ff67
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>