[442176] Stop using WorkspaceEPackageRegistry to detect models
XMIModelFileHandler does not check the namespaces present in the file
any more but just that the XMI namespace is present. Then we let the
resource load fail if an unknwon metamodel is found. 

We could have continued to check if the found namespace are present in
the EcorePackageRegistry (global or resource set one's) but with the
current behavior, the resource load will throws the expected errors.

Bug: 442176
Change-Id: Iac8de595dd13af4a63735699907b7de74aa98790
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>
Signed-off-by: Maxime Porhel <maxime.porhel@obeo.fr>