[443386] Reduce accesses to the preference store in GroupingContentProvider
The configurable options of GroupingContentProvider are stored in the
Eclipse preference store. Because they can change at (almost) arbitrary
time, the corresponding getters directly accessed the preference store
on each invocation to make sure they always get the current version.
However this code can be called a lot in deeply nested loops (by CNF),
and the store accesses are costly (compared to a raw field access). In
practice the values will not change very often (and almost certainly not
when we are inside a loop).

Replace theses accesses with direct field access from a separate
Configuration object, which uses a listener on the preference store to
keep the fields values up to date.

The DynamicConfigurationHelper is a first attempt at a generic helper
that could be used to fix these kinds of patterns elsewhere in the code,
but the preferences story in Eclipse is relatively complex (between
IPreferenceStore, IPreferenceService, IEclipsePreferences, and maybe
others), so it is kept internal fow now until we decide on a common
strategy.

Bug: 443386
Change-Id: If5817917527c7df2139628b94bfd4343748faddc
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>