[443286] Install the AirdCrossReferenceAdapter after the domain creation
Move the AirdCrossReferenceAdapterImpl after the domain creation.

In Bug 437646, AirdDCrossReferencerAdapter installation has been moved
in the session factory and a common adapter is installed on the resource
set (we add previously on adapter per aird and only on AirdResourceImpl
see Bug 437646). 

It was done just after the resource set creation (by resource set
factory which uses the first found extension) but just after the add of
the adapter, we create the editing domain: we call the
EditingDomainInFactoryService, which looks in the extension the 'first
most overrider' factory and create an editing domain. The creation 
receives the previously created set as parameter, but the created 
editing domain can use its own resource set. 

Our adapter installation was done too soon. This typically occurs with
Capella which does not use one editing domain per session, but has its
own shared editing domain.

Bug: 443286
Change-Id: If27099f4523c1a9c574ccaca7995d20f9c01dcf5
Signed-off-by: Maxime PORHEL <maxime.porhel@obeo.fr>