[437528] Fix wrong rectilinear bendpoints at opening.
* In some cases, the GMF bendpoints coordinates can be very far away
from the ones
displayed by Draw2D, especially when there are only two bendpoints(they
are always connected to the source and target figure, so the shift is
not visible). If the edge
centering property changes at opening and the edge is rectilinear, we
compute its rectilinear bounds from the GMF information (since we don't
have the figure yet) to center the last segment toward the node center.
This fix prevents wrong rectilinear bounds by
computing them from the GMF bounds of source and target nodes.

Bug: 437528
Change-Id: If6303994539eb896864ee7fc6ff6fd8db3af22da
Signed-off-by: Florian Barbin <florian.barbin@obeo.fr>