[437097] Handle case of drag'n'drop
* During drag'n'drop of a node, we store the existing layout of all
children of the draged node (and of its incoming and outgoing edges).
With this commit, the layout of incoming and outgoing edges are now
adapted to avoid to be "moved" after drop. This is done in
EdgeLayoutData.init(Edge).
* Part of code from
SetConnectionBendpointsAccordingToExtremityMoveCommmand has been extract
in adaptPointListAndRefPoints to be called during drag'n'drop.
* The layout stored at the begining of the execution of the tool is used
during the creation of the GMF view (in ConnectionsFactory for edges).
This code has been adapted because in case of dnd of border node, the
stored location of this border node and the associated point list can be
false because the border node is "moved" with the DBorderItemLocator.

This commit fixes the problem of comment 10 and 11 of bug 437097.

Bug: 437097
Change-Id: I44ac3c4dfd4c1a646a3df8be47c23dbdb8fe2d02
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>