[437528] Use Figure absolute bounds instead of GMF when possible.
* In the case where the edge EditPart is available, we now use the
source and target draw2D absolute bounds instead of GMF. Because of the
difference between the GMF absolute bounds computation and the draw2D
absolute bounds, the edge
centering behavior was wrong in the case of container within a
container.
* Create a new method to get the top level FreeformViewport to compute
the diagram scroll size and not the first container one.
* See https://bugs.eclipse.org/bugs/show_bug.cgi?id=437528#c9 for more
details.

Bug: 437528
Change-Id: I34ce80800c069d0dbddb5d8dfe0882694ae4fcd0
Signed-off-by: Florian Barbin <florian.barbin@obeo.fr>