[444759] Fix jump links status problem
If you enable the "specific default values" in connections preference
page and set it to "Rectilinear", all edges, created after that, have
the jump links set to "Above" instead of "None".

And also fix an activation problem at opening when the checkbox is checked (the combo is gray). This problem is here since the commit 9057209.

Bug: 444759
Change-Id: Idbaac8bf31bc93e149b9a535c07163fcf48e030b
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>