[437097] Handle case of drag'n'drop of source AND target
The commit a354dd83bf9b7ab23f1621b6242e5da438dd7310 handles the case of
drag'n'drop of one element but not the case when both source and target
have been dragged and dropped simultaneously.

*  A new method has been added in SiriusLayoutDataManager to retrieve
the opposite edge layout data of another edge layout data. An edge
layout data stored in the incomingEdgeLayoutDatas ref of its parent data
can have an opposite edge layout data in the outgoingEdgeLayoutDatas of
the other ends.
* The unused case CAUSED_BY_BOTH of EdgeLayoutData has been removed.
* The main changes are in ConnectionsFactory, that now handles the case
of drag'n'drop of source AND target.

Bug: 437097
Change-Id: Iea043150bd131a4a29fa0ec7b4cfa462c3e85f4c
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>