[430724] Introduce a ResourceMigrationMarker
This marker can be attached to a Resource to indicate that some
automatic migration *might* have changed the structure of the model
during the loading (without setting isModified=true).

The marker's presence forces the IsModifiedSavingPolicy to consider the
resource as "logically modified".

This marker is automatically removed when the resource is processed by
the saving policy which checked that the serialization has no difference
with the current file content.

Bug: 430724
Change-Id: I2364dfeb5ee5b7f8404033a49aba1a34ddc432a0
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>