[441483] New feature and update-site with all tests dependencies
Add a new feature named oes.tests and a new update-site
oes.sirius.tests.update, which for now contains only the sample
metamodels and modelers which are needed by the automated tests, but not
the tests themselves (they will be added in a later step).

Because some of the tests and test modelers depend on "Sirius Query
Legacy" which itself depend on Sirius, the new feature and update-site
must be built in a separate step, using augmented Target Platforms which
include both Sirius and Sirius Query Legacy. See the
releng/oes.targets/sirius_tests*.* files.

By default, these new Tests Target Platform get their version of Sirius
from the latest nightly. To get a full consistent build using a Sirius
core built using the exact same sources as the tests themselves, the
location where Sirius is obtained can be overridden using a Maven
settings file. This requires to jump through some hoops because Maven
settings require absolute paths, which we can not predict. The new
script build-all.sh at the root shows how this can be done.

The new update-site/repository is published in a sub-directory of the
main one, but the shortcut URLs (e.g. updates/nightly/latest/luna)
aggregate boths, so downstream consumers do not have to be aware of the
physical split.

Bug: 441483
Change-Id: I06edf372223c43026bbea2256680858d134b67a6
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>