[441483] Fix runtime errors in SWTBot test suite
The first test in the SWTBot suite,
ViewpointSpecificationProjectCreationTest, creates a plug-in project,
which triggers PDE. PDE requires PlatformAdmin to work properly, but
this was removed in Luna [1], and we did not have the workaround in
place to force the compatiblity bunde to be present.

Add the required dependency via the surefire plug-in configuration to
ensure everything needed is present at runtime (see [2]).

Also add a explicit dependencies to org.eclipse.sdk.ide and
org.eclipse.platform to the SWTBot tests plug-in, which seem to fix a
seemingly unrelated error about "Product org.eclipse.sdk.ide could not
be found." at the beginning of the tests.

[1] https://wiki.eclipse.org/Equinox/Luna_Framework#Removal_of_the_PlatformAdmin_Service_Implementation
[2] https://issues.jboss.org/browse/JBIDE-16161

Change-Id: I3661bba68cd66e9f8919624cde830c206266d207
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>