[test] Make SiriusControlAndDeleteRepresentationTest more reliable
This test directly use resource.save(SomeOutputstream) to snapshot in
memory models.
Calling resource.save() has side-effects on the resourceset notably
setting isModified and notifying all the listeners which lead to an
unexpected amount of logic triggered because of the test behavior.

This change inhibit those side effects to have a more reliable test.

Change-Id: I2528c70d618c98154d9e524a6decba86d212780d
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr>