[test] Fix exceptions thrown by CreateInstanceOperationTest and ChangeContextOperationTest
The tests tried to obtain the interpreter associated to an EObject
through an API which assumes the EObject is inside a Sirius session, but
there is no session created in these tests. Use a plain default
interpreter instead.

Note that the tests did not report any failure before, but created many
hundred lines of errors/stack traces in the log.

Change-Id: I83936fc22806e852e204aa51828b777550bb5a6f
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>