[cleanup] Only try to obtain an interpreter if we have an precondition to check
In AbstractRepresentationDialectServices.checkPrecondition(), only try
to obtain an interpreter associated to the semantic element if we
actually have a precondition expression to evaluate.

This fixes exceptions thrown during
DialectManagerTest.testDiagramDialectRegistration(), which calls
getAvailableRepresentationDescriptions() on "semantic elements" which
are not actually in a session. The proper fix would be to rewrite the
test to create such a session.

Change-Id: I3d2f06ecfaaecdaea707ee1d488d67773c94a79b