[444057] No missing bendpoint on target reconnection
Because the feedback is handled by SiriusConnectionEndPointEditPolicy
and the reconnection is handled by SiriusGraphicalNodeEditPolicy, when
the reconnection switches to another target candidate some bendpoints
can be missing on reconnection. The missing bendpoints are the result of
the ObliqueRouter that removes bendpoints over the target candidate.
+
Addition of a fix in ReconnectionCommandBuilder to avoid a  CCE with
EdgeMappingImport

Bug: 444057
Change-Id: Ib47c343a28989595d5af6157fa9e5ed1d658525a
Signed-off-by: Steve Monnier <steve.monnier@obeo.fr>