[439871] Add new headless target platform
Based on Luna only for now, it contains the minium required stuff to
build all the modules in the "headless" profile and produce the
oes.headless.update repository.

The new platform is used when the "headless" profile is enabled and the
"full" profile is not. From the command-line, a headless build can be
invoked with:

  % mvn -P!full clean package

Note that you may need to escape the "!" character with "\!" depending
on your shell.

Bug: 439871
Change-Id: Ide46b33393f7e3c6c57c12938153cb44a9eef293
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>