[442289] Improve gap computation
Currently, the gap is rounded to the lower Integer. So the last gap can
be different. To minimize the delta between the last gap and the others,
the gap is now rounded up to the lower Integer if the remainder is less
than or equal to 0.5 and higher integer if the remainder is greater than
0.5.

Bug: 442289
Change-Id: Ia1c234af77157abc6f61bf4aa28498ddb721f494
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>