[test] Fix the problem of 1 pixel delta for testResizingC2ToSWWith50()
Fix the problem of 1 pixel delta on some PC: "The point #0 is different
after resizing:  expected:<Point(229.0, 85.0)> but was:<Point(229.0,
86.0)>"

Cause of problem: At end of method
org.eclipse.gmf.runtime.draw2d.ui.internal.routers.RouterHelper.resetEndPointsToEdge(Connection,
PointList): sourceAnchorPoint is a PrecisionPoint with
preciseX=85.9985xxxx but was rounded to 85. After the resize of the node
the rounded value is 86 and it's "normal".

So the test has been adapted to have a delta of 1 pixel for this failing
case.

Change-Id: Iee75ba96d023c77ced07777d8337c3c0876cf039
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>