[test] Rename usecase data for this test
The data used for
RunRepairTest.testNoDirtyAfterMigrationWithFilterApplied has been
initially copied from CompositeFilterTest but not renamed.
The test RunRepairTest.testNoDirtyAfterMigrationWithFilterApplied fails
randomly on jenkins but not locally.
I think that this problem is due to a mix between the to use case data
without real explanation. 
This commit only renames the data to see if the problem persists or not
on jenkins.

Change-Id: I008e3b26f0314fa2a4a5afc91beea1963fd751fb
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>