[447963] Close the EditingSession before the Session is closed
This changes add a mechanism so that the EditingSession uses the
SessionListener.CLOSING notification to close its editors (while the
Session is still opened).

A RefreshJob which might have been triggered by a previous operation
will be waited for through a join() call so that it does not fail
because of the session being disposed.

Bug: 447963
Change-Id: I5cb536ee5c1b30f85c13ac0406afde8a52a0d393
Signed-off-by: Cedric Brun <cedric.brun@obeo.fr