[445603] Do not save in a transaction if the domain is disposed
If we can detect the Transaction Editing Domain has already been
disposed when doSave() is invoked, do not try to use an exclusive
transaction, as its commit() will fail with an NPE in this case, causing
the later parts of doSave() not to be executed even though the resources
may have actually been saved correctly on disk.

If we cannot know for sure when the save is invoked (as is the case in
SaveSessionJob which is called asynchronously), play it safe and disable
the use of a transaction unconditionally.

Bug: 445603
Change-Id: I83ca3122ee1f9d84c1ee56aee532137b44f85cb8
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>