[test] The waited popup for session opening has not the right name.
After the opening of a session (UIResource.openSession();), we wait
that the load progress monitor popup is closed. But the popup title
currently used, "Progress Information", is not the right one. Currently,
the title is "Loading models" locally and "Loading models (as username)"
on Jenkins.
This probably explains some random failing tests (for example
ExportDiagramAsImagesTest).

Change-Id: Ie70d6afe977677ba2ec9af801a5b9e94e479d3ad
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>