[test] Use the default SWTBotPreferences.SCREENSHOTS_DIR folder
Use the default SWTBotPreferences.SCREENSHOTS_DIR folder for
screenshots. This allows to override it by using system property
<org.eclipse.swtbot.screenshots.dir> as explained in the javadoc of this
constant. This is useful under Tycho to generate screenshots in the
"target" folder to be cleaned at each launch.

After this commit the pom.xml that launches test will be updated to
override <org.eclipse.swtbot.screenshots.dir> property with a line like
this:

<org.eclipse.swtbot.screenshots.dir>${project.build.directory}/screenshots</org.eclipse.swtbot.screenshots.dir>

Change-Id: If6baaf093572dcaee11d78722405f892740cb8aa
Signed-off-by: Laurent Redor <laurent.redor@obeo.fr>