[422490] Add the basicfamily sample
This includes the metamodel (with generated implementation, edit and
editor plug-ins), a small sample model, and a sample Sirius modeler with
very basic capabilities. It corresponds to the material needed to follow
the online tutorial (currently hosted on the wiki).

All these plug-ins are built during the normal build, but only to
validate they compile. They are actually packaged in the
oes.sample.family plug-in (contributed by the oes.samples feature) as
examples that can be imported as source in the user's workspace, but are
not installable directly.

Bug: 422490
Change-Id: If84c6934a2384ae3fe75dfc68aa0bb29bc245684
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>