[430724] Only enable the new saving policy if XText resources are present
Instead of enabling the new saving policy for all sessions as soon as
the Sirius/Xtext bridge plug-in is installed, enable it only when at
least one Xtext resource is added. We do not revert to the default one
if all Xtext resources are removed, but this should not happen very
often.

Note that with this change, even the Xtext-specific saving options are
now only enabled if Xtext resources are present, not just the new saving
policy.

Bug: 430724
Change-Id: Ie13112a6468243e6bce24beab1c32026dcc969a5
Signed-off-by: Pierre-Charles David <pierre-charles.david@obeo.fr>